Turn CSP report only on for all apps that use GOV.UK CSP
This follows on from the previous commit which removed the global
environment variable from the configmap and now applies this variable to
each application that uses it individually.

I have wrapped the true in quotes to ensure it is kept as a string and
not parsed as a YAML boolean

This environment variable sets the GOV.UK Content Security Policy to
report only mode.

This has been applied to:

- collections
- draft-collections
- email-alert-frontend
- feedback
- finder-frontend
- frontend
- draft-frontend
- government-frontend
- draft-government-frontend
- info-frontend
- licencefinder
- signon (lone backend app with CSP)
- smartanswers