Turn off "report-only" for CSP for UXM owned apps
This change begins the process of turning the GOV.UK Content Security
Policy on for the apps on https://www.gov.uk. This starts this process
of initially applying this change just to apps owned by the User
Experience Measurement (UXM) team to see if how it goes before applying
to the rest of the frontend apps.

As we won't learn much from this until this is in the production
environment I've set this to apply to all environments.