Fix `Map keys must be unique` errors in VS Code.
YAML [mappings] are not allowed to contain duplicate keys, but some
parsers treat the `<<` key (the [merge key]) as a special case and not
only allow duplicates but will perform all of the merges (as opposed to
rejecting the whole document as not valid, or dropping all but one of
the duplicates). In other words, we may have been relying on undefined
or at least non-standard behaviour.

The sigs.k8s.io/yaml parser used in Helm (and pretty much all Kubernetes
tooling) safely handles multiple `<<` by performing all of the merges,
but some parsers (such as the one in VS Code) complain about duplicates
of the `<<` key.

Example of non-conformant YAML which causes VS Code to complain:

```yaml
annotations:
  <<: *default-alb-ingress-annotations
  <<: *alb-ingress-www-waf-ruleset
```

The more portable alternatives are:

```yaml
annotations:
  <<: [*default-alb-ingress-annotations, *alb-ingress-www-waf-ruleset]
```

or:

```yaml
annotations:
  <<:
    - *default-alb-ingress-annotations
    - *alb-ingress-www-waf-ruleset
```

Further reading:

[mappings]: https://yaml.org/spec/1.2.2/#mapping
[merge key]: https://yaml.org/type/merge.html

Tested: zero diff in Helm output; all the expected LB annotations are
present with and without this change.