Fix lints in ruletests + remove broken playbook link.
The playbook link in the alert description was making the alert tests
fail and with the way it's currently implemented (in the Prometheus
template rather than in the Alertmanager template) it's not
straightforward to fix the test to support or even just to make the test
ignore it.

So for now, we'll live without the playbook since having working tests
is more important.

The way to reinstate the playbook link is to use an Alertmanager
template instead of a Prometheus template. That'll also make it work
automatically for all alerts without needing to add the `playbook` label
in every alert rule. I'll have a go at that in a followup.