Give whitehall-admin workers more RAM headroom in prod.
https://github.com/alphagov/whitehall/pull/7963 helped smooth out the
memory usage but there's at least a suspicion that there are still some
cases lurking that cause spikes of large allocations that can OOM
workers during publishing tasks, which (because of poor design that
we've inherited) can cause tasks to be lost and not retried.

So let's increase the limits by 50% to reduce the chance of being bitten
by this, while we fix the underlying issue.