Improve the Slack failure notification for post-sync.
Most people deploying code to GOV.UK probably don't care about the Argo
CD lifecycle terminology, so "post-sync" isn't super helpful as a name.

Change the wording of the message to convey what the workflow does.

Also refer to it as post-deploy rather than post-sync, since in the
developer-facing tooling we mostly talk about deployment rather than
sync (i.e. the Argo CD term). I'm less sure about this part though,
because the workflow is internally still called "post-sync". OTOH it's
still just as easy to search for the error message 🤷

Kudos to @chao-xian for pointing out the wtf-ness of the message :D