Integration assets ingress needs to be public.
Same as production already is. It's only staging that's supposed to be
GDS-only.

This just restores the same access that people had before we moved to
Kubernetes, when we mistakenly gave the assets LB the wrong firewall
ruleset.

The reason it needs to be publicly reachable is the long-standing
oddball usage of the integration environment for training users of the
content management system (which doesn't necessarily take place at GDS).

This is acceptable for now (at least until we can run the end-to-end
stack in minikube/kind/k3s), because draft assets are still behind
auth-proxy and there should never be anything sensitive in integration
anyway (unlike staging, which is allowed to contain pre-publication
content from production).