Let whitehall-admin use memcache.
I suspect this will help with the 0.1% 504 rate on whitehall-admin.

https://govuk.zendesk.com/agent/tickets/5434688

This memcache has < 2% CPU utilisation, < 25% network utilisation and >
14 GiB of available RAM, so adding whitehall-admin should be small beer.

The name is somewhat misleading: it's frontend in the normal sense
rather than the oddball GOV.UK sense, i.e. it's a memcache for apps that
have UIs, not just apps that serve pages to the general public. Other
GOV.UK "backend" apps (i.e. frontends for content management tools) use
this memcache already.