Add shield advanced protection to ALBs
We had this enabled on the load balancers for our EC2 instances, and it's
worth bringing across. There's a few benefits, including proactive engagement
from the AWS SRT team and better monitoring options.

We already pay for the subscription as part of the org, so it's free to us
at this point.