Bump link-checker-api to v23
Something seems to have gone wrong with this workflow. Now GitHub
actions is happy to trigger Argo Workflows, but Argo Workflows doesn't
seem to be able to do the deploy:

[Example failed Argo Workflow](https://argo-workflows.eks.production.govuk.digital/workflows/apps/link-checker-api-deploy-image-integrationzjd9g?tab=workflow&nodeId=link-checker-api-deploy-image-integrationzjd9g&nodePanelView=summary)

I've checked ECR and the v23 tag is already there, so I think we just
need to bump the helm config to get this going again.