Delete existing branch before updating image tag
This prevents failures on update image tag workflow re-runs. This is
because subsequent runs attempt to re-use the existing branch which can
be out of date, cause subsquent pushed to be rejected.

We should just always start from a clean branch, as we don't care what
previous workflow runs have done.