Replace transition rake task to import orgs only
The Publishing Platform team has added support in the Transition app for
adding and deleting sites and hosts directly, through a GUI, thus
allowing us to retire transition-config.

The rake task `import:all:orgs_sites_hosts` was used to import:
- organisations from whitehall
- sites and hosts from transition-config
into transition.

In the new way of doing things, the source of truth for sites and hosts
is going to be the transition app itself.

Organisations still need to be imported from whitehall.