Include mean and max (not sum) in summary tables
sum (called "Total" in the Grafana UI) isn't meaningful for Total HTTP
Request Errors per second - it's the sum of the per second error rate
multiplied by the number of rate intervals on the graph, not the number of
seconds. This is why it comes up with odd statistics like support having
had a total of 0.4 errors over the last 6 hours. What that probably means
is that for one rate interval, support had an error rate of 0.4 errors /
second.

Best to be consistent and show mean (average) everywhere I think. I've
also added "max", because I find this useulf to sort by.