Fix calculation of Average Request Duration
Previously, the expression for average request duration was incorrect
(or at least confusing).

http_request_duration_seconds was a prometheus summary, giving quantiles
(e.g. the median, or the 95th percentile) for individual apps and
individual pods. Averaging these quantiles by job doesn't entirely make
sense - if a job has multiple pods, those pods will have different
medians, and avg() is going to take the mean across pods of the median
request durations. That's almost certainly not what we want.

Prometheus histograms track the number of observations and the sum of
the observed values, which allows you to calculate the average by
dividing the sum by the count. Which is what we're doing here.