Fix variable values
http_request_duration_seconds was a prometheus summary. We switched to
prometheus histograms in https://github.com/alphagov/govuk_app_config/pull/318

The histogram metric is called http_request_duration_seconds_bucket
(note the extra `_bucket`). Now that most of the apps don't use the
http_request_duration_seconds metric, it's missing most of the app
labels, making it the wrong place to query to get the list of apps.

Instead, the http_requests_total metric has everything we need.

I've also removed the "quantile" variable, as we're no longer using the
summary metric which includes quantile as a label.