Switch to heatmap for request durations
The old Request duration per App/Controller/Action panel was very noisy,
and possibly inaccurate.

Without an app / quantile specified (i.e. by default) it would try to
show 2860 series (one series for every combination of app, controller,
action _and_ quantile).

Even with an app and quantile specified, some apps (e.g. whitehall) have
a large number of controllers and actions, so the graph would still be
unusably noisy.

And even if you managed to find an interesting series, the query itself
is still the average (mean) across pods of a number of summaries
(medians, 95th percentiles etc.), which isn't a very meaningful
statistic.

I think it would be more useful to use the histograms (which we have
now, thanks to https://github.com/alphagov/govuk_app_config/pull/318) to
show a heatmap.

Heatmaps show the change in distribution over time, so they give us a
visual indication both of how many requests are happening, and how many
fall into each request duration bucket. In this case, by default we'll
bundle all the requests to all the apps together, and allow segmenting
by app using the variable on the dashboard. It is possible (and useful)
to segment further (down to the controller / action), but we're probably
trying to do too much with a single dashboard by trying that.