Use `.*` as the value for all
The expressions on this dashboard use the `=~` operator to match a
regular expression of job. By default, grafana constructs a string like
this:

    (asset-manager|bouncer|collections|collections-publisher|contacts-admin|content-publisher|content-store-postgresql-branch|content-tagger|draft-collections|draft-content-store-postgresql-branch|draft-frontend|draft-government-frontend|draft-router-api|draft-static|email-alert-api|feedback|finder-frontend|frontend|government-frontend|info-frontend|link-checker-api|manuals-publisher|maslow|publisher|router-api|search-admin|search-api|service-manual-publisher|short-url-manager|smartanswers|specialist-publisher|static|support|whitehall-admin)

which works fine, but it's a bit verbose (and mayyybbeee not that
performant). Since we know we're always using the app variable in `=~`
conditions, we can simplify it by using `.*`.