Move transition-config cron job to Transition app
* Remove custom transition-config-cron and
* replace it with a standard cron task in the Transition app's config.

transition-config-cron made data from the transition-config repository
available to the Transition app and then invoked one of the Transition
app's Rake tasks.

The Rake task that it was invoking:

1. already lives in the Transition app, and
2. no longer requires data from the transition-config repository (since
   https://github.com/alphagov/govuk-helm-charts/pull/1296).

This change is already live in integration
(https://github.com/alphagov/govuk-helm-charts/pull/1310), now this
makes the same change for staging and production, too.