Stop using transition-config-cron in integration
And instead, replace it with a standard cron task for the Transition app.

If I've understood transition-config-cron correctly, it makes data from
the transition-config repository available to the Transition app and
then invokes one of the Transition app's Rake tasks.

The Rake task that it invokes:

1. already lives in the Transition app, and
2. no longer requires data from the transition-config repository (since
   https://github.com/alphagov/govuk-helm-charts/pull/1296)

I'm having trouble fully understanding the transition-config-cron file,
so I'm testing the waters first by limiting this change to integration.
If it works as expected, I'll then make this change for production and
staging and remove transition-config-cron altogether.