Use WITH (FORCE) when restoring Postgres databases.
This is the least brittle/evil way to avoid problems with locking (e.g.
`database ... is being accessed by other users` doing a full restore.

The old govuk_env_sync jobs used to kick out existing clients and set
the connection limit to zero (which of course could end up being left
that way if the job were to fail).

Tested: ran a restore in staging using this.