Ignore CSV preview pages
The URL structure of these pages cause an issue where only either the
csv file or the preview can be saved. This is becuase for the preview
pages the crawler creates a directory that is the same name of the CSV
file (e.g. "example.com/media/foo.csv"). If the directory is created
first this prevents the file from being saved as the directory with same
name already exists, and vice versa. This behaviour is also present in
the old mirror script. For now to ensure we have a mirror of all the csv
files, let ignore the preview.