Add last updated file to monitor the mirror freshness
This file can be used to check the freshness of mirrors by observing the
"Last-Modified" header. This adds a mirror only file that is updated
everytime the mirror sync runs successfully.

We have to use the s5cmd "cp" command directly as the "sync" will only
upload files where the file size has changed.