Configure backups for Postgres databases.
This replaces
https://github.com/alphagov/govuk-puppet/tree/main/modules/govuk_env_sync
for most of our Postgres databases, i.e. daily full backups to S3 and
copying into the staging/integration environments (which doubles as an
automated restore test).

I've kept the timings roughly similar to the old Puppet-managed jobs,
but with the minutes randomised to help reduce the usage spikes.

There are a few jobs which aren't supported yet (MySQL and jobs which
involve transforming the data), but I've included them commented out for
the sake of being able to see the whole picture. They'll be enabled
shortly.