Record 100% trace data for Whitehall integration
This overrides the sample rate that's set for all apps in
`charts/app-config/templates/env-configmap.yaml`. It controls how much
data OpenTelemetry sends to Grafana Tempo.

We're building a Service Level Indicator for publishing
(Whitehall->Content Store) and experimenting with traces in Tempo to
figure out whether that's the right tool for the job. Since our
OpenTelemetry sample rate is currently set to 0.1%, there isn't actually
much data in Tempo to experiment with.

The issue isn't just the lack of traces of interest, either. I've
repeatedly run into an issue where a relevant trace that we've recorded
is linking to another that's missing... I can't definitely say why, but
because 0.1% is such a low number, I wonder if this might fix that, too.

We're only making this change in integration for now. If Tempo looks
like a good fit for the SLI, I think this same setting or at least
a high sample rate would be needed in production too.