Make yamllint line length rules less annoying.
- Make line-length an error rather than a warning. (Warnings are
  annoying because we don't enforce them but they still nag, so they get
  merged and then everyone gets nagged forever.)
- Allow lines to break the limit where it's unavoidable.

It's my bad that this setup was a bit wonky to start with; certainly
didn't realise I'd implicitly removed `allow-non-breakable-words` from
the defaults.

This way we avoid having lint rules that we don't enforce (i.e. just
noise) and still keep the readability benefits of the line length
heuristic without false positives.