Remove disused MQ monitoring heartbeat.
This used to drive a non-paging/dashboard-only Icinga alert that would
show when a heartbeat message hadn't been read from the queue for > 5
minutes.

We now have a much better MQ dashboard so the heartbeat cronjob is
obsolete. Hadn't spotted that when we ported this over; assumed it was
some sort of keepalive workaround or something (otherwise we'd have
dropped it months ago).