Transform Signon app URLs appropriately in staging.
There used to be a horrible hack in the govuk_env_sync script[1] that
would substitute the prod domain name for the staging domain name when
running the nightly restore of the prod Signon database into Staging.

The new backup/restore scripts don't do this (since it was brittle and
complicated and has nothing to do with backup/restore), so do the
substitution in the Signon config. This is way simpler to understand and
keeps the config where it belongs, along with the rest of the staging
configuration for Signon.

[1]:
https://github.com/alphagov/govuk-puppet/blob/aa75027/modules/govuk_env_sync/files/govuk_env_sync.sh#L537-L573