Reduce gzip compression on backups.
gzip (and pg_restore) set the zlib compression level to 6 by default,
which takes several times the CPU time of level 1.

The backup jobs are bottlenecked on CPU (and are single-threaded), and
we don't care about a extra few % improvement in compression ratio.

This should save cores and speed up backups (and possibly also
restores, to a lesser extent).