Use resources-finalizer in staging/production.
We sensibly took a precautionary approach with this to begin with, but
it turns out that not having the Argo CD resources-finalizer clean up
when an app object is removed is quite problematic: after an app is
removed, we end up with disused workloads taking up resources and
orphaned PDBs which interfere with cluster upgrades.

Cascading deletion of owned resources on app deletion turns out to be
the lesser of evils.

We don't currently have any persistent storage managed through this
template, so there's no problem there. If we do start introducing
persistent storage then we'd have to consider the object lifecycle
carefully anyway — and we have options there.