Increase limits for Asset Manager virus scanning
We are seeing issues where users uploading some files have those marked
as "infected" when they are not.

In the two cases that we have investigated, this is caused by ClamAV
raising either `Heuristics.Limits.Exceeded.MaxFiles` or
`Heuristics.Limits.Exceeded.MaxScanSize`.

Therefore increasing the size limit for MaxFiles from the default of
10,000 to 20,000 and MaxScanSize from 1200M to 2000M.

These new limits have been tested on a local version of ClamAV and have
confirmed the affected files are now properly scanned.