Add Sentry configuration for Asset Manager worker
We are not currently seeing errors from Asset Manager workers being
logged to Sentry.

This is because there is no configuration of the SENTRY_DSN environment
variable, which is configured for all other GOV.UK workers.

Therefore adding this so we can see errors which occur within the worker
process (including virus scanning).