Rename database more robustly on restore.
Let's just put up with having two code paths depending on whether the
target database already exist or not.

Alternatives considered:

Considered putting the logic in a PL/pgSQL procedure[1], but that seemed
like it'd be harder to troubleshoot if anything unexpected happens
later. As much as I dislike that the shell script is getting more
complex, at least it's easy enough to see what what's gone wrong from
the trace when it fails.

[1]: https://github.com/alphagov/govuk-helm-charts/pull/1451#discussion_r1394116850