content-data workers spool entire reports in RAM :(
Give them more RAM for now, at least so we can see what's going on more
easily (and hopefully get it going again in the short term).

We also don't need 3 workers cos these sit idle like 99.9% of the time
and the Sidekiq tasks are huge and not sharded so more workers doesn't
make it go faster.