Set pipefail only where we really want it.
The only place we actually want pipefail is where we stream the backup
to S3. If the producer (pg_dump/mysqldump/mongodump) terminates with an
error then we want to consider the whole pipeline failed.

Experience has shown that in the restore case, we're better off leaving
it to the consumer (at the end of the pipeline) to determine the overall
exit status.