Cron job for govuk-sli-collector
I've mostly cribbed this from the templates for
`content-store-mongo-to-postgres-cron` and `dependabot-metrics`.

_Re: the cron schedule string_

The `govuk-sli-collector` script needs to know how often it's being run,
so I want to run this job every `INTERVAL_MINUTES` minutes and therefore
need to keep the variable and the cron schedule in sync. I was going to
just add an inline comment about it, but I'd prefer this string
interpolation approach if it actually works.

I did worry that we might one day change `intervalMinutes` to a bad
value, leaving the cron job running on an unexpected schedule or not
running at all. But I'd like to think that we'd notice in either case.
(And if it does run but the value isn't a valid integer, the
`govuk-sli-collector` script itself with complain.)