Don't try to terminate rdsadmin connections.
Amazon's RDS flavor of Postgres has some irritating quirks regarding the
superuser attribute. It's not possible to log in as a superuser or to
assign the superuser attribute to a role, and the replacement
`rds_superuser` role doesn't work very much like an actual superuser.

So we need to be more selective about which connections we try to
terminate before swapping databases around at the end of a full restore,
because Amazon doesn't let us grant ourselves sufficient privileges to
terminate all connections.

This makes it a bit easier to avoid rolling back where we shouldn't,
while also not masking real errors that we want to surface.

Tested: installed via Helm in staging and successfully ran a restore of content-data-admin.