Consolidate L7 LBs in production.
Same as 9141aeb (#1519) but for production.

Put most of the "backend" apps (i.e. the content management system web
UIs plus a couple of Internet-facing APIs) behind a single (regional set
of) AWS ALBs.

This is essentially the same load balancer configuration we used to run
before the switch to Kubernetes (except without the brittle Terraform
list/count horror show that used to leave production completely hosed if
you dared remove a decommissioned backend from the list).

If I've not fluffed the arithmetic, this saves:
- 72 routeable IPv4 addresses (billed at 5¢/hr starting Feb 2024)
- 24 ALBs (at 2.52¢/hr)