[1/4] Temporarily stop publishing-api updating content-store in production
Re-do of #1499 for switching over the live content-store.
To make sure that we have a completely up-to-date content-store database at
the point we switchover the proxy to PostgreSQL, we need to temporarily
disable publishing-api workers while the mongo-to-postgresql job runs one
last time.

It also adds the `ENQUEUE_PUBLISH_INTENTS=true` env var, which will
make propagation of publish intents asynchronous (once the [publishing-api
PR](https://github.com/alphagov/publishing-api/pull/2566) is merged)

[Trello card](https://trello.com/c/kQbZFteC/954-prepare-advance-prs-for-swapping-the-live-content-store-in-production)

I'm raising this as a draft PR so that it doesn't get accidentally merged
before the scheduled maintenance period, will convert to a regular PR at
that time.