Remove custom sentry DSN secret name for integration live content-store
After removing the proxy, we found that the `content-store` app would not
start, with config errors relating to secret names not being found.

The custom secret names for rails secret key base and sentry dsn were
added to allow dual-running of two content-stores, but they don't appear
to work with a single app that's using a different repo name.

Removing the custom name worked for the rails secret key base (#1603)
This PR does the same for sentry DSN