Switch cron job from integration to production
We've got `govuk-sli-collector` working in integration and now we need
to run it in production so that it can access real data and do the job
it's intended for.

We're removing it from integration now because we believe that the costs
of running there it outweigh the benefits (of having it readily
available for future testing).