Reinstate image-tags for govuk-sli-collector job
I've mostly copied this from the `govuk-application` template.

We recently extracted this standalone chart out of the deprecated
`govuk-jobs` chart. Since that change meant no longer using the
`govuk-application` template, we lost the ability for the cronjob to
pull arbitrary images from ECR using `image-tags`. It can now only pull
the `latest`-tagged (i.e. `main` branch) image.

After some deliberation, I've decided to add the use of `image-tags` to
this standalone chart, allowing us to deploy arbitrary images again. The
main motivation for this is for ease/flexibility of future development.
It's hard to ignore that significant changes could benefit from the
ability to deploy feature branches to integration. And while temporarily
enabling the cronjob in integration will require only a small change,
re-enabling the use of arbitrary images will be harder work that isn't
guaranteed to be approachable for whoever is tasked with making the
application code changes.

`automatic_deploys_enabled` and `promote_deployment` are set to false
because (as I understand it) they don't mean anything for this
non-`app-config` chart.