Switch production draft content-store back to the content-store ECR repo
We removed the proxy and the mongo content-store applications yesterday,
we've also removed the legacy EC2 mongo instances, so we are ready to
switch the content-store applications back to the `content-store` ECR
repo in production.

Similar PRs in integration and staging went through with zero disruption
so we're expecting the same for prod.