Use environment-specific image for cronjob
Since this isn't an app, it doesn't use Argo workflows or `image-tags`,
but we'd still like to be able to deploy separate images to each
environment as and when it's needed. We've made some corresponding
changes to make this possible in [`govuk-infrastructure`](https://github.com/alphagov/govuk-infrastructure/pull/1046)
and [`govuk-sli-collector`](https://github.com/alphagov/govuk-sli-collector/pull/6).

This cronjob only currently runs in production, but I can easily imagine
wanting to temporarily re-enable it in integration for some testing,
debugging, etc. When that happens, I don't want the developer who's
given that task to have to figure this other stuff out, I want them to
be able to just add the integration-equivalent version of this
`values-production.yml` entry into `values-integration.yml`.