Enable deployment via image tag files for govuk-sli-collector
This adds the image tag files for govuk-sli-collector and references
them from the argo-app. This allows us to use the existing deployment
process to deploy the govuk-sli-collector.

The govuk-sli-collector is only deployed into the production
environment. However, the image tags files are available incase it's
needed to be deployed into either integration or staging for testing.