Slow down MongoDB statefulset rollouts in nonprod.
Also remove the dead replicas from the initial config on the command
line in integration (otherwise it's not a representative test).

Alternatives considered:

Startup/liveness/readiness probes would be a more complete solution, but
only if they're bug-free; otherwise it's likely to introduce more
problems than it solves.

Ideally we'd use an operator (such as mongodb-kubernetes-operator) to
handle the edge cases, but Mongo 2.6 is too old for that.

So, since this is really only meant to be a temporary installation until
we get rid of router/router-api altogether, let's see how we get on with
a crude but (hopefully) effective delay in the rolling update.