Suppress connection logspew from router-mongo.
The [manual] recommends against this, but the logspew is really quite
excessive and makes it difficult to see anything else that's going on
without filtering. We're also paying to store and index this stuff.

It's also trivial to turn verbose logging back on at runtime as needed:

```js
db.adminCommand({setParameter: 1, quiet: 0});
```

[manual]: https://www.mongodb.com/docs/v2.6/reference/configuration-options/#systemlog-options