Add govuk_content_production back into backup config.
It turns out this MongoDB/DocumentDB database was only *mostly* owned by
content-store (which was migrated to Postgres). It's still accessed
by publisher, specialist-publisher and service-manual-publisher.

Need to tease out what on earth is going on with these three apps and
figure out how bad the situation is and how to fix it, but that's a
different matter. For now, just keep backing this stuff up so at least
we're not SOL if we lose prod.