Increase router nginx's client_max_body_size for licensify requests
We allow files upto 10MB to be uploaded to licensify, however
since moving to EKS we've nginx at router has server 413 responses to
users. The reason we're only seeing it since the migration, is that
previously file upload request bypassed router completely. They were
uploaded on a seperate `uploadlicence` domain that went directly to
licensify frontend.