Stop copying nonprod databases around on weekends.
Little if anything changes over the weekend, so there's really no reason
to keep running these restore jobs on Saturday and Sunday mornings. Some
of them are nontrivial in terms of resource usage.

Leaving the integration Signon job as-is because integration Signon is
independent from prod/staging, so it's still (somewhat disturbingly)
worth keeping daily full backups of it.