Add placeholder image tag file for licensify apps
The post sync workflow currently fails for the licensify-frontend,
licensify-feed and licensify-admin apps. This is becuase the workflow
expects the image tag file to match the ArgoCD App name. However this is
"licensify" for all three apps, becuase they are all deployments within
the same ArgoCD App. This will supress the deployment failures, but
doesn't matter as we want it to be a no-op anyway.