Scale up `search-api-v2` replicas
This increases the replicas for `search-api-v2` to be in line with those
for `search-api` in anticipation of us going "fully" live shortly.

This can/should be reevaluated after a few days with full traffic to see
if it needs adjusting - we may not need quite as much capacity as v1.