Remove info-frontend
Info pages have been unpublished and thus info-frontend isn't serving
any traffic regarding info pages.

We are seeing the odd request for legacy /needs pages [1] (~1 a day)
which are being received by info-frontend. info-frontend doesn't know
about these pages and so a 404 is returned and so they do not exist on
the public facing website and only exist in the publishing stack
(published by Maslow). After we remove info-frontend, these requests
will return a 503, as the backend app serving the needs pages will no
longer be available.

Returning 503s isn't ideal but we agreed [2] to is the best approach given:
- the traffic is tiny
- the traffic mainly comes from crawlers
- we will only return 503s up until we retire Maslow and then we'll
  return a normal 404

The alternative would be to unpublish the 679 editions / routes now in
our publishing stack but we run the risk of strange behaviour / errors
as we have inconsistent states between our apps.

Trello:
https://trello.com/c/NcSjgmmq/2385-remove-from-govuk-helm-charts

[1]: https://kibana.logit.io/s/13d1a0b1-f54f-407b-a4e5-f53ba653fac3/goto/9f68346dc881bd4913aaeaf185522608?security_tenant=global
[2]: https://gds.slack.com/archives/C03D4A72HGE/p1709568749883129?thread_ts=1709556305.963639&cid=C03D4A72HGE